PDO code examples updates.
