+++ examples/PDO/advanced.example.php
@@ -20,10 +20,12 @@ use Ranking\AlgorithmInterface;
 use Ranking\PDO\Advanced as AdvancedAlgorithm;
 use Ranking\Mysql\Object;
 
+$condition = '%s%';
 // Select a random name to search for.
 $sql = <<<SQL
 SELECT name
   FROM $table
+  WHERE `name` LIKE '$condition'
  ORDER BY RAND() ASC
  LIMIT 1
 SQL;
@@ -34,36 +36,18 @@ $name = $row['name'];
 
 $total_time = microtime(true);
 try {
-    $ranking_obj = new AdvancedAlgorithm($pdo, $table, $row_score, $rank_row);
+    $ranking_obj = new AdvancedAlgorithm($pdo, $table, $row_score);
+
+    $ranking_obj->addColumnCondition('name', $condition, 'LIKE');
+
+    $ranking_obj->addAltOrderByColumn('name');
 
     if (!$ranking_obj->isReady()) {
         $ranking_obj->run();
     }
 
-    $object = new Object($ranking_obj);
-
-    $object->setAttributes(array('name' => $name));
-
-    $rank = $object->getRank();
-
-    $score = $object->getScore();
-
-    print 'Rank of ' . $table . ' row with ' . $data_row . ' = ' . $name .
-        ' is : ' . $rank . "\n";
-    print 'Score is ' . $score . "\n\n";
-    // List 5 rows before and 5 following the current rank.
-    $rank -= 6;
-    if ($rank < 0) {
-        $rank = 0;
-    }
-    $names = array_map(function ($model) use (&$rank, $data_row, $row_score) {
-        $arr = $model->getAttributes();
-        return ++$rank . ' : ' . $arr[$data_row] . ' : ' . $arr[$row_score];
-    }, $ranking_obj->getRowsAtRank($rank, 11));
-    print count($names) . ' rows of ' . $table . ' starting at rank = ' . ($rank - count($names) + 1) .
-        ' is :' . "\n";
-    print 'Rank        name       Score' . "\n";
-    print implode("\n", $names) . "\n";
+    print "Only searching at names containing 's' \n";
+    print_my_rank($ranking_obj, $table, $data_row, $row_score, $name);
 
 } catch (Exception $e) {
     print $e->getMessage() . "\n";
