+++ examples/advanced.example.php
@@ -8,9 +8,7 @@
 
 require_once 'settings.php';
 
-use Ranking\AlgorithmInterface;
 use Ranking\Mysql\AdvancedRanking;
-use Ranking\Mysql\Object;
 
 $condition = '%s%';
 // Select a random name to search for.
@@ -53,31 +51,3 @@ $mysqli->close();
 
 echo "Took " . (microtime(true) - $total_time) . "s to complete\n";
 
-function print_my_rank(AlgorithmInterface $ranking_obj, $table, $data_row, $row_score, $name)
-{
-
-    $object = new Object($ranking_obj);
-
-    $object->setAttributes(array('name' => $name));
-
-    $rank = $object->getRank();
-
-    $score = $object->getScore();
-
-    print 'Rank of ' . $table . ' row with ' . $data_row . ' = ' . $name .
-        ' is : ' . $rank . "\n";
-    print 'Score is ' . $score . "\n\n";
-    // List 5 rows before and 5 following the current rank.
-    $rank -= 6;
-    if ($rank < 0) {
-        $rank = 0;
-    }
-    $names = array_map(function ($model) use (&$rank, $data_row, $row_score) {
-        $arr = $model->getAttributes();
-        return ++$rank . ' : ' . $arr[$data_row] . ' : ' . $arr[$row_score];
-    }, $ranking_obj->getRowsAtRank($rank, 10));
-    print count($names) . ' rows of ' . $table . ' starting at rank = ' . ($rank - count($names) + 1) .
-        ' is :' . "\n";
-    print 'Rank        name       Score' . "\n";
-    print implode("\n", $names) . "\n";
-}
